moved json directory loader logic to json loader. Also, added a way to add from a json file
